+++ GRDB/FetchRequest/SQLCollation.swift
@@ -46,8 +46,8 @@ extension _SQLCollatedExpression : _SQLOrdering {
     /// Do not use it directly.
     ///
     /// See https://github.com/groue/GRDB.swift/#the-query-interface
-    public func orderingSQL(db: Database, inout _ bindings: [DatabaseValueConvertible?]) throws -> String {
-        return try sqlExpression.orderingSQL(db, &bindings)
+    public func orderingSQL(db: Database, inout _ arguments: StatementArguments) throws -> String {
+        return try sqlExpression.orderingSQL(db, &arguments)
     }
 }